requirements.txt
fastapi
uvicorn
pandas
scikit-learn
google-genai
python-dotenv
joblib
streamlit
requests